04 - RAG/requirements.txt
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
openai==1.36.0
python-dotenv==1.0.1
langserve==0.2.2
sse_starlette==2.1.2
langchain_chroma==0.1.2
sentence-transformers==2.6.1